+++ actions.go
@@ -17,8 +17,31 @@ func RegisterAction(eventName string, action func(string) (string, error)) {
 	registeredActions[eventName] = append(registeredActions[eventName], action)
 }
 
-func EmitActions(event string, message string, sender User, room Room, connector chan<- Event) {
-	connector <- NewEvent(event, message, nil, sender, room)
+func EmitActions(event string, message string, sender User, room Room, bot Bot, connector chan<- Event) {
+	connector <- NewEvent(event, message, nil, sender, room, bot)
+}
+
+func EmitDirectedMessageActions(message string, sender User, room Room, bot Bot, connector chan<- Event) {
+	EmitActions(EventDirectedMessage, message, sender, room, bot, connector)
+}
+
+func EmitChannelMessageActions(message string, sender User, room Room, bot Bot, connector chan<- Event) {
+	EmitActions(EventChannelMessage, message, sender, room, bot, connector)
+}
+
+func EmitMessageActions(message string, sender User, room Room, bot Bot, connector chan<- Event) {
+	regex := bot.DirectedMessageRegex()
+	result := regex.FindAllStringSubmatch(message, -1)
+
+	if len(result) > 0 {
+		EmitDirectedMessageActions(result[0][1], sender, room, bot, connector)
+	} else {
+		EmitChannelMessageActions(message, sender, room, bot, connector)
+	}
+}
+
+func EmitTickActions(message string, sender User, room Room, bot Bot, connector chan<- Event) {
+	EmitActions(EventTick, message, sender, room, bot, connector)
 }
 
 func actionWorker(toActions <-chan Event, toConnector chan<- Event) {
@@ -38,11 +61,11 @@ func actionWorker(toActions <-chan Event, toConnector chan<- Event) {
 			}
 
 			actionFound = true
-			toConnector <- NewEvent(event.Type(), result, err, event.Sender(), event.Room())
+			toConnector <- NewEvent(event.Type(), result, err, event.Sender(), event.Room(), event.Bot())
 		}
 
 		if !actionFound {
-			toConnector <- NewEvent(event.Type(), "", ErrNoActionFound, event.Sender(), event.Room())
+			toConnector <- NewEvent(event.Type(), "", ErrNoActionFound, event.Sender(), event.Room(), event.Bot())
 		}
 	}
 }