+++ events.go
@@ -1,19 +1,30 @@
 package dosbot
 
+const (
+	EventDirectedMessage = "directed-message"
+	EventChannelMessage  = "channel-message"
+	EventTick            = "tick"
+)
+
 type Event struct {
 	eventType string
+	bot       Bot
 	msg       string
 	err       error
 	sender    User
 	room      Room
 }
 
-func NewEvent(eventType string, msg string, err error, sender User, room Room) Event {
-	return Event{eventType: eventType, msg: msg, err: err, sender: sender, room: room}
+func NewEvent(eventType string, msg string, err error, sender User, room Room, bot Bot) Event {
+	return Event{eventType: eventType, bot: bot, msg: msg, err: err, sender: sender, room: room}
 }
 
 func (e Event) Type() string {
 	return e.eventType
+}
+
+func (e Event) Bot() Bot {
+	return e.bot
 }
 
 func (e Event) Message() string {